% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{prefs_schema}
\alias{prefs_schema}
\title{Look up RStudio Preference Schema}
\usage{
prefs_schema(version = NULL, quiet = FALSE)
}
\arguments{
\item{version}{The version of the RStudio IDE, e.g. \code{"1.3.959"},
\code{"1.4.1717"}, or \code{"2021.09.0+351"}.}

\item{quiet}{Suppress console messages and output}
}
\description{
Finds and downloads the RStudio user preference schema for your current
version of the IDE (or the most recent released version of the IDE). The
preference details for all released versions of the IDE are included in the
\pkg{rsprefs} package. The preferences from the latest version are available
in \link{prefs_rstudio}, but note that those preferences may not match your
version of RStudio unless you're using the latest preview version available
at \url{https://dailies.rstudio.com}.
}
\examples{
prefs_schema("2021.09.0+351")

}
