% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefs.R
\name{rs_prefs_reset_defaults}
\alias{rs_prefs_reset_defaults}
\title{Reset Preferences to the RStudio Default}
\usage{
rs_prefs_reset_defaults(source = c("project", "user"), verbose = FALSE, ...)
}
\arguments{
\item{source}{The source of the current preference value. Preferences are set
at different levels, from lowest to highest precedence:
\itemize{
\item \code{"default"} are RStudio's built-in defaults
\item \code{"computed"} are detected or supplied from external sources, e.g. the
path to \code{git} on your system
\item \code{"system"} are derived from system-wide \code{rstudio-prefs.json}
\item \code{"user"} are set by the user for all sessions (global options)
\item \code{"project"} are set by the current project session
}

The default is \code{"user"}, since these are the settings you set yourself for
all projects. You can include any number of sources, or \code{"all"} to include
all preferences regardless of source.}

\item{verbose}{Prints or suppress informative output}

\item{...}{
  Arguments passed on to \code{\link[=rs_prefs_snapshot]{rs_prefs_snapshot}}
  \describe{
    \item{\code{include}}{Names of RStudio preferences to include. If provided, only
these preferences are included. See \link{rstudio_prefs} for preference names.}
    \item{\code{exclude}}{Names of RStudio preferences to exclude from the snapshot. See
\link{rstudio_prefs} for all of the preference names.}
    \item{\code{exclude_os_prefs}}{Excludes operating-system or machine-dependent
system preferences from the snapshot or the snapshot restore.}
  }}
}
\value{
Resets preferences to their built-in defaults, returning the current
preferences invisibly as a list. You can also return to the preferences
prior to \code{rs_prefs_reset_defaults()} with \code{\link[=rs_prefs_snapshot_undo]{rs_prefs_snapshot_undo()}}.
}
\description{
Returns currently set preferences to RStudio's original, built-in default.
}
\examples{
\dontrun{
rs_prefs_reset_defaults()
}

}
