% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefs.R
\name{rs_prefs_snapshot}
\alias{rs_prefs_snapshot}
\alias{rs_prefs_snapshot_apply}
\alias{rs_prefs_snapshot_list}
\title{Snapshot RStudio Preferences}
\usage{
rs_prefs_snapshot(
  name,
  path = NULL,
  include = NULL,
  exclude = NULL,
  source = "user",
  overwrite = FALSE
)

rs_prefs_snapshot_apply(name = NULL, path = NULL, verbose = FALSE)

rs_prefs_snapshot_list(path = NULL, verbose = TRUE)
}
\arguments{
\item{name}{The name of the snapshot to save or apply.
\itemize{
\item In \code{rs_prefs_snapshot()}, set to \code{NULL} to show the preferences that
would be included in the snapshot.
\item In \code{rs_prefs_snapshot_apply()}, set to \code{NULL} to list available snapshots.
}}

\item{path}{A GitHub gist ID or local path where the snapshot should be
saved. To create a new public gist, set \code{path = "new gist"}. For a new
private gist, use \code{path = "new private gist"}.}

\item{include}{Names of RStudio preferences to include. If provided, only
these preferences are included. See \link{rstudio_prefs} for preference names.}

\item{exclude}{Names of RStudio preferences to exclude from the snapshot. See
\link{rstudio_prefs} for all of the preference names.}

\item{source}{The source of the current preference value. Preferences are set
at different levels, from lowest to highest precedence:
\itemize{
\item \code{"default"} are RStudio's built-in defaults
\item \code{"computed"} are detected or supplied from external sources, e.g. the
path to \code{git} on your system
\item \code{"system"} are derived from system-wide \code{rstudio-prefs.json}
\item \code{"user"} are set by the user for all sessions (global options)
\item \code{"project"} are set by the current project session
}

The default is \code{"user"}, since these are the settings you set yourself for
all projects. You can include any number of sources, or \code{"all"} to include
all preferences regardless of source.}

\item{overwrite}{If the snapshot exists, should it be overwritten?}

\item{verbose}{Prints or suppress informative output}
}
\description{
Save named snapshots of your RStudio preferences that you can apply later.
\itemize{
\item \code{rs_prefs_snapshot()}: Save your RStudio user preferences with a name that
you can later use to identify these particular settings.
\item \code{rs_prefs_snapshot_apply()}: Apply a saved snapshot.
\item \code{rs_prefs_snapshot_list()}: List available snapshots.
}
}
\examples{
if (interactive()) {
  tmpfile <- tempfile(fileext = ".json")
  rs_prefs_snapshot("example", path = tmpfile)
  rs_prefs_snapshot_list(tmpfile)
}

}
