% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_prefs.R
\docType{data}
\name{rstudio_prefs_v}
\alias{rstudio_prefs_v}
\title{rstudio_prefs_v}
\format{
A list with preferences from 7 released
versions of RStudio.
}
\source{
\url{https://github.com/rstudio/rstudio/raw/main/src/cpp/session/resources/schema/user-prefs-schema.json}
}
\usage{
rstudio_prefs_v
}
\description{
The RStudio User Preferences, their description and default value for
released versions of RStudio.
}
\seealso{
\link{rstudio_prefs}
}
\keyword{datasets}
