% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefs.R
\name{snapshot_prefs}
\alias{snapshot_prefs}
\alias{snapshot_prefs_save}
\alias{snapshot_prefs_use}
\alias{snapshot_prefs_list}
\alias{snapshot_prefs_undo}
\title{Snapshot RStudio Preferences}
\usage{
snapshot_prefs_save(
  name,
  path = NULL,
  include = NULL,
  exclude = NULL,
  source = "user",
  exclude_os_prefs = TRUE,
  overwrite = FALSE,
  preview = FALSE
)

snapshot_prefs_use(
  name = NULL,
  path = NULL,
  exclude_os_prefs = TRUE,
  verbose = FALSE,
  preview = FALSE
)

snapshot_prefs_list(path = NULL, verbose = TRUE)

snapshot_prefs_undo(verbose = TRUE)
}
\arguments{
\item{name}{The name of the snapshot to save or apply.
\itemize{
\item In \code{snapshot_prefs_save()}, set to \code{NULL} to show the preferences that
would be included in the snapshot.
\item In \code{snapshot_prefs_use()}, set to \code{NULL} to list available snapshots.
}}

\item{path}{A GitHub gist ID or local path where the snapshot should be
saved. To create a new public gist, set \code{path = "new gist"}. For a new
private gist, use \code{path = "new private gist"}.}

\item{include}{Names of RStudio preferences to include. If provided, only
these preferences are included. See \link{prefs_rstudio} for preference names.}

\item{exclude}{Names of RStudio preferences to exclude from the snapshot. See
\link{prefs_rstudio} for all of the preference names.}

\item{source}{The source of the current preference value. Preferences are set
at different levels, from lowest to highest precedence:
\itemize{
\item \code{"default"} are RStudio's built-in defaults
\item \code{"computed"} are detected or supplied from external sources, e.g. the
path to \code{git} on your system
\item \code{"system"} are derived from system-wide \code{rstudio-prefs.json}
\item \code{"user"} are set by the user for all sessions (global options)
\item \code{"project"} are set by the current project session
}

The default is \code{"user"}, since these are the settings you set yourself for
all projects. You can include any number of sources, or \code{"all"} to include
all preferences regardless of source.}

\item{exclude_os_prefs}{Excludes operating-system or machine-dependent
system preferences from the snapshot or the snapshot restore.}

\item{overwrite}{If the snapshot exists, should it be overwritten?}

\item{preview}{When \code{TRUE}, previews the preferences that will be or are
included in the snapshot, but does not save or apply them.}

\item{verbose}{Prints or suppress informative output}
}
\description{
Save named snapshots of your RStudio preferences that you can apply later.
\itemize{
\item \code{snapshot_prefs_save()}: Save your RStudio user preferences with a name that
you can later use to identify these particular settings.
\item \code{snapshot_prefs_use()}: Apply a saved snapshot.
\item \code{snapshot_prefs_list()}: List available snapshots.
\item \code{snapshot_prefs_undo()}: Undo the last applied snapshot.
}
}
\examples{
if (interactive()) {
  tmpfile <- tempfile(fileext = ".json")
  snapshot_prefs_save("example", path = tmpfile)
  snapshot_prefs_list(tmpfile)
}

}
